requirements.txt
asgiref==3.2.7
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.5
django-appconf==1.0.4
django-compressor==2.4
django-libsass==0.8
django-rest-auth==0.9.5
django-rq==2.3.2
django-sass==1.0.0
django-sass-processor==0.8
django-static-precompiler==1.8.2
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
libsass==0.20.0
Pillow==7.1.2
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
python-dotenv==0.13.0
pytz==2020.1
rcssmin==1.0.6
redis==3.5.2
reportlab==3.5.42
requests==2.23.0
rjsmin==1.1.0
rq==1.4.1
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
xhtml2pdf==0.2.4
